fix(api): move response schemas to api
Ref. metriport/metriport-internal#1333
