+++ packages/api/src/domain/medical/ihe-result.ts
@@ -1,8 +1,4 @@
-import {
-  PatientDiscoveryResponse,
-  DocumentQueryResponse,
-  DocumentRetrievalResponse,
-} from "@metriport/ihe-gateway-sdk";
+import { z } from "zod";
 import { BaseDomainCreate } from "../base-domain";
 
 export interface BaseResultDomain extends BaseDomainCreate {
@@ -11,14 +7,37 @@ export interface BaseResultDomain extends BaseDomainCreate {
   createdAt: Date;
 }
 
-export interface PatientDiscoveryResult extends BaseResultDomain {
-  data: PatientDiscoveryResponse;
-}
+export const issue = z.object({
+  severity: z.string(),
+  code: z.string(),
+  details: z.object({ text: z.string() }),
+});
 
-export interface DocumentQueryResult extends BaseResultDomain {
-  data: DocumentQueryResponse;
-}
+export const operationOutcome = z.object({
+  resourceType: z.string(),
+  id: z.string(),
+  issue: z.array(issue),
+});
 
-export interface DocumentRetrievalResult extends BaseResultDomain {
-  data: DocumentRetrievalResponse;
-}
+export const baseResponseSchema = z.object({
+  id: z.string(),
+  cxId: z.string(),
+  timestamp: z.string(),
+  responseTimestamp: z.string(),
+  xcpdPatientId: z.object({ id: z.string(), system: z.string() }).optional(),
+  patientId: z.string(),
+  operationOutcome: operationOutcome.nullish(),
+});
+
+export const documentReference = z.object({
+  homeCommunityId: z.string(),
+  docUniqueId: z.string(),
+  repositoryUniqueId: z.string(),
+  contentType: z.string().nullish(),
+  language: z.string().nullish(),
+  uri: z.string().nullish(),
+  creation: z.string().nullish(),
+  title: z.string().nullish(),
+});
+
+export type DocumentReference = z.infer<typeof documentReference>;
