+++ packages/ihe-gateway-sdk/src/models/document-retrieval.ts
@@ -1,6 +1,4 @@
-import { BaseRequest, DocumentReference, baseResponseSchema, documentReference } from "./shared";
-
-import { z } from "zod";
+import { BaseRequest, DocumentReference } from "./shared";
 
 export type DocumentRetrievalRequest = BaseRequest & {
   gateway: {
@@ -10,16 +8,3 @@ export type DocumentRetrievalRequest = BaseRequest & {
   patientId: string;
   documentReference: DocumentReference[];
 };
-
-export const docFileReference = documentReference.extend({
-  newRepositoryUniqueId: z.string(),
-  newDocUniqueId: z.string(),
-  url: z.string(),
-});
-
-export const documentRetrievalResponseSchema = baseResponseSchema.extend({
-  documentReference: z.array(docFileReference),
-  gateway: z.object({ homeCommunityId: z.string(), url: z.string() }),
-});
-
-export type DocumentRetrievalResponse = z.infer<typeof documentRetrievalResponseSchema>;
