+++ packages/ihe-gateway-sdk/src/models/shared.ts
@@ -47,25 +47,3 @@ export const documentReference = z.object({
 });
 
 export type DocumentReference = z.infer<typeof documentReference>;
-
-export const issue = z.object({
-  severity: z.string(),
-  code: z.string(),
-  details: z.object({ text: z.string() }),
-});
-
-export const operationOutcome = z.object({
-  resourceType: z.string(),
-  id: z.string(),
-  issue: z.array(issue),
-});
-
-export const baseResponseSchema = z.object({
-  id: z.string(),
-  cxId: z.string(),
-  timestamp: z.string(),
-  responseTimestamp: z.string(),
-  xcpdPatientId: z.object({ id: z.string(), system: z.string() }).optional(),
-  patientId: z.string(),
-  operationOutcome: operationOutcome.nullish(),
-});
